Fix automatic-time wait argument type

+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/broker/mam/TestCase2516571.kt
@@ -187,7 +187,7 @@ class TestCase2516571 : AbstractMsalUiTest(){
             }
 
             ThreadUtils.sleepSafely(
-                TimeUnit.SECONDS.toMillis(1),
+                TimeUnit.SECONDS.toMillis(1).toInt(),
                 "waiting for automatic time",
                 "interrupted while waiting for automatic time"
             )
